Fix duplicities in proposed suggestions
In case, when you are writing slowly search input value, in combination with slow async suggestion provider, method updateQuerySuggestions is triggered multiple times (previous execution is not yet finished). Result is that suggestions are duplicated. 

Proposed solution will wait for condition: inputValue is the same as searchInputValue and skip the state, when termToSuggestFrom is empty. I don't understand exactly, why condition !termToSuggestFrom has been in line 168.